% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffVar.R
\name{coeffVar}
\alias{coeffVar}
\title{Calculates the coefficient of variation}
\usage{
coeffVar(x)
}
\arguments{
\item{x}{a vector}
}
\value{
Coefficient of variation, 100 * SD / mean.
}
\description{
Calculates the coefficient of variation (as a percent)
}

