% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rland.R
\name{getParamsRandom}
\alias{getParamsRandom}
\title{Gets random parameters without desired characteristics}
\usage{
getParamsRandom(reps)
}
\arguments{
\item{reps}{number of parameter sets to generate}
}
\value{
matrix of control parameters with columns
        n1, n2, n3, n4, hAsp, pHM, nHoriz, nVert, children, spread, inh
}
\description{
Arbitrary bounds that should usually work
}

