% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAreas.R
\name{makeAreas}
\alias{makeAreas}
\title{Takes a summary matrix (from class landscape), a subset with info on the "good" polygons.
Calculates areas.}
\usage{
makeAreas(summ)
}
\arguments{
\item{summ}{a summary matrix (from class landscape)}
}
\value{
A matrix with columns index, area, x, and y, one row per "good" polygon
}
\description{
Takes a summary matrix (from class landscape), a subset with info on the "good" polygons.
Calculates areas.
}

