\name{makeTargetMat}
\alias{makeTargetMat}
\title{Generates specific targets from target ranges}
\usage{
  makeTargetMat(targets, tryGuess)
}
\arguments{
  \item{targets}{a named list of target ranges for each
  characteristic}

  \item{tryGuess}{the number of targets to generate}
}
\value{
  A matrix of specific targets. Column order "n",
  "degMean", "degSD", "areaCV", "hAsp"
}
\description{
  None
}

