% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlandGui.R
\name{rlandGui}
\alias{rlandGui}
\title{Shiny app for rland}
\usage{
rlandGui()
}
\description{
Provides a GUI interface for creating landscapes. There are two tabs,
"Single Run" and "Batch Mode".
}
\details{
The Single Run tab is essentially an interface to \link{rlandscape}. 
It lets you explore parameters for creating a single landscape.

Batch Mode is an interface to \link{rland}, where you set targets
and bounds for output landscape characteristics and generate
many landscapes with characteristics within those bounds.

Both interfaces provided by the GUI are simplifications - they
omit the detailed control parameters. To control those you will
need to use \code{rlandscape} or \code{rland} directly.
}
\seealso{
\code{\link{rlandscape}} for a single landscape,
\code{\link{rland}} for batches of landscapes.
}

