% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlandscape-package.R
\docType{package}
\name{rlandscape-package}
\alias{rlandscape-package}
\title{rlandscape: Random Landscape Generation for Harvest Scheduling Models}
\description{
With the \code{rlandscape} function you can generate individual landscapes with
explicit control over the control parameters (e.g. hole proportion, merge proportion,
number of points placed with each method) used to create the landscape.
The \code{rland} function is designed for batch productions where the desired
characteristics of the resulting landscapes (e.g. degree mean, area coefficient
of variation) are specified. Rland uses previously run regressions to determine
suitable control parameters. Rland can also be used through a graphical user interface (GUI).
This can be started by entering \code{rland.gui()} on the R console.
code{\link{rlandscape}}, \code{\link{rland}}, \code{\link{plot.landscape}}
}
\references{
Gregor Passolt, Miranda J. Fix, and Sandor F. Toth. A Voronoi
Tesselation-based Approach to Generate Hypothetical Forest Landscapes.
\emph{Canadian Journal of Forest Research}, 2012.
\url{http://www.nrcresearchpress.com/doi/abs/10.1139/cjfr-2012-0265}
 (2012pks
}

