% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveAdj.R
\name{saveAdj}
\alias{saveAdj}
\title{Saves an adjacency list}
\usage{
saveAdj(adj, filename)
}
\arguments{
\item{adj}{list of adjacencies}

\item{filename}{character, filename (and path) to save, will have "_adj.txt" appended.}
}
\value{
No return.
}
\description{
Saves an adjacency list
}

