\name{saveAreas}
\alias{saveAreas}
\title{Saves an area list}
\usage{
  saveAreas(keepers, filename)
}
\arguments{
  \item{keepers}{data.frame with 4 columns, index, area, x
  and y.}

  \item{filename}{character, filename (and path) to save
  to.}
}
\value{
  Coefficient of variation, 100 * SD / mean.
}
\description{
  Saves an area list
}

