\name{makeAdjList}
\alias{makeAdjList}
\title{Creates an adjacency list}
\usage{
  makeAdjList(edges, n, merges = NULL)
}
\arguments{
  \item{edges}{a data.frame with edge data}

  \item{n}{the number of kept polygons}

  \item{merges}{=merges that occurred}
}
\value{
  A list of adjacencies
}
\description{
  Creates an adjacency list
}

