\name{makeAreas}
\alias{makeAreas}
\title{Takes a summary matrix (from class landscape), a subset with info on the "good" polygons.}
\usage{
  makeAreas(summ)
}
\arguments{
  \item{summ}{a summary matrix (from class landscape)}
}
\value{
  A matrix with columns index, area, x, and y, one row per
  "good" polygon
}
\description{
  Calculates the coefficient of variation (as a percent)
}

