\docType{package}
\name{rlandscape-package}
\alias{rlandscape-package}
\title{rlandscape: Random Landscape Generation for Harvest Scheduling Models}
\description{
  With the \code{rlandscape} function you can generate
  individual landscapes with explicit control over the
  control parameters (e.g. hole proportion, merge
  proportion, number of points placed with each method)
  used to create the landscape. The \code{rland} function
  is designed for batch productions where the desired
  characteristics of the resulting landscapes (e.g. degree
  mean, area coefficient of variation) are specified. Rland
  uses previously run regressions to determine suitable
  control parameters. Rland can also be used through a
  graphical user interface (GUI). This can be started by
  entering \code{rland.gui()} on the R console.
  code{\link{rlandscape}}, \code{\link{rland}},
  \code{\link{plot.landscape}}
}
\references{
  Gregor Passolt, Miranda J. Fix, and Sandor F. Toth
  (2012).
}

